requirements.txt
alembic==1.0.11
amqp==2.5.1
anyjson==0.3.3
atomicwrites==1.3.0
attrs==19.1.0
billiard==3.6.1.0
celery==4.4.0rc3
certifi==2019.6.16
chardet==3.0.4
Click==7.0
coverage==4.5.3
Django==2.2.3
django-appconf==1.0.3
django-celery==3.3.0
django-compressor==2.3
django-haystack==2.8.1
django-mdeditor==0.1.14
django-redis==4.10.0
django-redis-sessions==0.6.1
django-secret-key==1.0.1
django-tinymce==2.8.0
django-uuslug==1.1.8
dnspython==1.16.0
eventlet==0.25.1
fdfs-client-py==1.2.6
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
# Editable Git install with no remote (flaskr==1.0.0)
-e /Users/mac/Documents/projects/python_self_teaching/flask-tutorial
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.18
itsdangerous==1.1.0
jieba==0.39
Jinja2==2.10.1
kombu==4.6.4
Mako==1.0.14
MarkupSafe==1.1.1
mistune==0.8.4
monotonic==1.5
more-itertools==7.2.0
mutagen==1.42.0
mysqlclient==1.4.2.post1
packaging==19.0
Pillow==6.1.0
pluggy==0.12.0
py==1.8.0
pycryptodomex==3.7.2
Pygments==2.4.2
PyMySQL==0.9.3
pyparsing==2.4.1.1
pytest==5.0.1
python-alipay-sdk==1.10.1
python-dateutil==2.8.0
python-editor==1.0.4
python-memcached==1.59
python-slugify==3.0.2
pytz==2019.1
qiniu==7.2.6
rcssmin==1.0.6
redis==3.2.1
requests==2.22.0
rjsmin==1.1.0
six==1.12.0
SQLAlchemy==1.3.6
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.3
vine==1.3.0
waitress==1.3.0
wcwidth==0.1.7
Werkzeug==0.15.5
Whoosh==2.7.4
WTForms==2.2.1
zipp==0.5.2
